Rebalanced item drops. Removed Cobblestone "booby prize" and added new default drop functionality. Will require more balancing.
Fixed a double-call problem where Harvest was getting called twice. Can't
let the default harvest calls happen due to inability to allow the block
conversion to tile to handle that properly. Could update the block to
allow for this but that could be messy and prone to error. For now just
ignoring willHarvest, doing the harvest directly and returning false from
removedByPlayer.