Made minimum world gen for common ore 2 (max 5).
Added experience for mining.

Added Kerogen and Shale as a fuel source.

Increased base chance for Metamorphic chunks to drop.

Changed the gold shard to give a gold nugget instead of an ingot.

Added coal to SAG mill for Kerogen.

Added flint as a possible outcome for SAG mill on Shale.